pyproject.toml
[project]
name = "basesplit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "langchain>=0.3.15",
    "langchain-anthropic>=0.3.4",
    "langchain-openai>=0.3.2",
    "openai>=1.60.1",
    "plivo>=4.58.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
]
